moved more functions
